gorten/pkg/logs/logs.go
package logs

import (
	"github.com/sirupsen/logrus"
)

var Logger *logrus.Logger

func InitLogger() *logrus.Logger {
	logger := logrus.New()
	logger.SetLevel(logrus.DebugLevel)
	logger.SetReportCaller(true)

	logger.SetFormatter(&logrus.JSONFormatter{
		PrettyPrint:       true,
		DisableHTMLEscape: true,
		FieldMap: logrus.FieldMap{
			logrus.FieldKeyTime:  "@timestamp",
			logrus.FieldKeyLevel: "severity",
			logrus.FieldKeyMsg:   "message",
		},
	})

	//Inject logger into the Logger global
	Logger = logger

	return Logger
}
